Added appropriate status code (Updates parts3a.md)
Although a 400 status code is not an unsuitable fit for the case when a field or parameter is missing from a request body, a 400 Bad Request is a general error indicating that the request was malformed in some way, while a 422 (Unprocessable Entity) status code is more specific and indicates that the server understood the request but the request body was invalid or malformed. 422 is more specific and speaks directly to the issue. It is hard to make a false guess when the server responds with a 422.